quên mật khẩu bằng otp

+++ datn/app/Http/Controllers/Client/WishlistController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Product;
 use App\Models\User;
+use Carbon\Carbon;
 use App\Models\Wishlist;
 use Illuminate\Support\Facades\Auth;
 
@@ -13,6 +14,7 @@ class WishlistController extends Controller
 {
     //
     public function add_wishlist($id){
+        $this->authorize('member');
         $new_wishlist = new Wishlist();
         $new_wishlist->product_id = $id;
         $new_wishlist->user_id = Auth::user()->id;
@@ -21,12 +23,15 @@ class WishlistController extends Controller
     }
 
     public function index(){
+        $this->authorize('member');
+        $today = Carbon::now('Asia/Ho_Chi_Minh')->format('H:i:s');
         $id_user = Auth::user()->id;
         $wish_lists = Wishlist::all()->where('user_id',$id_user);
-        return view('client.wishlist.index', compact('wish_lists'));
+        return view('client.wishlist.index', compact('wish_lists','today'));
     }
 
     public function remove_wishlist($id){
+        $this->authorize('member');
         $id_destroy = Wishlist::find($id);
         $id_destroy->delete();
